backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
google-generativeai
python-dotenv
